Move internal/x/strs to internal/strs
