+++ internal/x/lint/check_enum_zero_values_invalid.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/strs"
+	"github.com/uber/prototool/internal/strs"
 	"github.com/uber/prototool/internal/x/text"
 )
