+++ internal/x/lint/check_message_field_names_lowercase.go
@@ -22,7 +22,7 @@ package lint
 
 import (
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/strs"
+	"github.com/uber/prototool/internal/strs"
 	"github.com/uber/prototool/internal/x/text"
 )
