+++ internal/x/lint/check_rpc_names_camel_case.go
@@ -22,7 +22,7 @@ package lint
 
 import (
 	"github.com/emicklei/proto"
-	"github.com/uber/prototool/internal/x/strs"
+	"github.com/uber/prototool/internal/strs"
 	"github.com/uber/prototool/internal/x/text"
 )
